Organized the code into packages Also fixed colorgather as it wasn't being picked up as a .java file
