+++ Group6FinalProject/bin/.gitignore
@@ -1,0 +1,13 @@
+/CollectionSystem.class
+/ColorGather
+/LCDInfo.class
+/Map.class
+/Navigation.class
+/Odometer.class
+/RobotController.class
+/SensorMotor.class
+/USGather.class
+/navigation
+/collection
+/sensors
+/robot
