--- Group6FinalProject/src/HelloWorld.java
@@ -1,6 +1,0 @@
-
-public class HelloWorld {
-	public static void main(String[] args){
-		System.out.println("Hello World!");
-	}
-}
